Change type from int to size_t to eliminate warning

+++ util.cpp
@@ -23,14 +23,14 @@ void putline(std::ostream &out,const std::string &line)
 
 std::string sprintf( const char *fmt, ... )
 {
-    int size = strlen(fmt) * 3;   // guess at size
+    size_t size = strlen(fmt) * 3;   // guess at size
     std::string str;
     va_list ap;
     for(;;)
     {
         str.resize(size);
         va_start(ap, fmt);
-        int n = vsnprintf((char *)str.data(), size, fmt, ap);
+        size_t n = vsnprintf((char *)str.data(), size, fmt, ap);
         va_end(ap);
         if( n>-1 && n<size )    // are we done yet?
         {
